requirements.txt
scrapy
openai
python-telegram-bot
Pillow
dateparser
requests
beautifulsoup4
python-dotenv